// src/app/model/producto-model.ts
export class ProductoModel{
    id: number = 0;
    nombre: string = '';
    material: string = '';
    tipo: string = '';
    precio: number = 0;
    peso: number = 0;
    largo: number = 0;
    ancho: number = 0;
    stock: number = 0;
    descripcion: string = '';
    marca: string = '';
    talla: number = 0;
}